requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
discord==1.7.3
discord.py==1.7.3
google-auth==2.2.1
google-auth-oauthlib==0.4.6
gspread==4.0.1
httplib2==0.19.1
idna==3.2
multidict==5.1.0
oauth2client==4.1.3
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
